pyproject.toml
[project]
name = "gopro-transfer"
version = "0.1.0"
description = "A tool for transferring and managing GoPro media files."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "pydantic>=2.5.0",
    "python-dotenv>=1.0.0",
    "loguru>=0.7.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/gopro_transfer"]

[project.scripts]
gopro-transfer = "gopro_transfer.main:main"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_functions = "test_*"

[tool.ruff]
line-length = 89
target-version = "py310"
